requirements.txt
uvicorn
pymongo
fastapi
google-generativeai
sentence-transformers
python-dotenv
faiss-cpu
streamlit
scikit-learn